+++ Assets/Danmokou/Plugins/Danmokou/Danmaku/Math/MathRepos/VTPRepo.cs
@@ -150,8 +150,17 @@ public static class VTPControllers {
     public static ExVTP Velocity(ExCoordF cf) => (vel, dt, bpi, delta) => InLetCtx(vel, bpi, tac =>
         cf(vel.cos, vel.sin, tac, (x, y, z) =>
             Ex.Block(
+                //TODO: automatically detect if flipX/Y are present on custom data,
+                // if they are, use FVelocity instead.
                 delta.x.Is(vel.flipX.Mul(x).Mul(dt)),
                 delta.y.Is(vel.flipY.Mul(y).Mul(dt))
+            )));
+    
+    public static ExVTP FVelocity(ExCoordF cf) => (vel, dt, bpi, delta) => InLetCtx(vel, bpi, tac =>
+        cf(vel.cos, vel.sin, tac, (x, y, z) =>
+            Ex.Block(
+                delta.x.Is(ReflectEx.ReferenceExpr<float>(FiringCtx.FLIPX, tac).Mul(x).Mul(dt)),
+                delta.y.Is(ReflectEx.ReferenceExpr<float>(FiringCtx.FLIPY, tac).Mul(y).Mul(dt))
             )));
 
     public static ExVTP Velocity3D(ExCoordF cf) => (vel, dt, bpi, delta) => InLetCtx(vel, bpi, tac =>
@@ -168,6 +177,16 @@ public static class VTPControllers {
             delta.x *= vel.flipX * dT;
             delta.y *= vel.flipY * dT;
         };
+
+    public static VTP FVelocity(CoordF coordF) {
+        var kx = FiringCtx.GetKey(FiringCtx.FLIPX);
+        var ky = FiringCtx.GetKey(FiringCtx.FLIPY);
+        return delegate(ref Movement vel, in float dT, ref ParametricInfo bpi, ref Vector3 delta) {
+            coordF(vel.cos_rot, vel.sin_rot, bpi, ref delta);
+            delta.x *= (bpi.ctx.boundFloats.TryGetValue(kx, out var fx) ? fx : 1) * dT;
+            delta.y *= (bpi.ctx.boundFloats.TryGetValue(ky, out var fy) ? fy : 1) * dT;
+        };
+    }
 
     public static ExVTP Offset(ExCoordF cf) => (vel, dt, bpi, delta) => InLetCtx(vel, bpi, tac =>
         cf(vel.cos, vel.sin, tac, (x, y, z) =>